# bookster/src/main/java/com/gorgona/bookster/feedback/FeedbackResponse.java
package com.gorgona.bookster.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;

}